requirements.txt
asgiref==3.6.0
Django==4.2.1
django-formtools==2.4
sqlparse==0.4.4
tzdata==2023.3
